[SecurityBundle] Add AutowiringTypesTest
