+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/app/AutowiringTypes/config.yml
@@ -8,10 +8,3 @@ services:
 framework:
     templating:
         engines: ['php']
-security:
-    providers:
-        dummy:
-            memory: ~
-    firewalls:
-        dummy:
-            security: false
